+++ src/components/Popover/Popover.spec.js
@@ -1,22 +1,17 @@
 import React from "react";
-import Popover from "./Popover";
-import Badge from "../Badge/Badge.jsx";
 import renderer from "react-test-renderer";
-
-import { configure, mount, shallow } from "enzyme";
-import Adapter from "enzyme-adapter-react-16";
-
-configure({ adapter: new Adapter() });
+import Popover from "./Popover";
 
 describe("Default Popover", () => {
   const defaultExample = (
-    <Popover title="title" content="content">
+    <Popover title="title" body="content">
       <span>trigger</span>
     </Popover>
   );
 
   test("popover snapshop", () => {
-    const component = shallow(defaultExample);
-    expect(component).toMatchSnapshot();
+    const component = renderer.create(defaultExample);
+    const tree = component.toJSON();
+    expect(tree).toMatchSnapshot();
   });
 });