+++ src/components/Table/components/Search.jsx
@@ -9,7 +9,7 @@ const Search = props => {
   };
 
   return (
-    <div className="table-search ml-5">
+    <div className="table-search ms-5">
       <Form onSubmit={e => e.preventDefault()}>
         <Form.Control
           type="text"